+++ jest.config.js
@@ -1,0 +1,6 @@
+// Jest configuration as per https://jestjs.io/docs/configuration
+
+module.exports = {
+    globalSetup: './test/__global-e2e-setup.js',
+    globalTeardown: './test/__global-e2e-teardown.js',
+}